+++ tests/pytest_tests/system_tests/conftest.py
@@ -5,6 +5,7 @@ import platform
 import secrets
 import string
 import subprocess
+import time
 import unittest.mock
 import urllib.parse
 from collections.abc import Sequence
@@ -28,28 +29,15 @@ try:
 except ImportError:
     from typing_extensions import Literal
 
-from .testcontainer_setup_utils import (
-    DEFAULT_SERVER_CONTAINER_NAME,
-    DEFAULT_SERVER_URL,
-    DEFAULT_SERVER_VOLUME,
-    FIXTURE_SERVICE_PORT,
-    LOCAL_BASE_PORT,
-    SERVICES_API_PORT,
-    WandbServerSettings,
-    spin_wandb_server,
-)
 
-# `local-testcontainer2` url and ports
-# Normally these env vars are only set in CI
-default_server_url2 = (
-    os.getenv("WANDB_TEST_SERVER_URL2") if os.getenv("CI") else "http://localhost"
-)
-local_base_port2 = LOCAL_BASE_PORT if os.getenv("CI") else "9180"
-service_api_port2 = SERVICES_API_PORT if os.getenv("CI") else "9183"
-fixture_service_port2 = FIXTURE_SERVICE_PORT if os.getenv("CI") else "9115"
+# `local-testcontainer` url and ports
+DEFAULT_SERVER_URL = "http://localhost"
+LOCAL_BASE_PORT = "8080"
+SERVICES_API_PORT = "8083"
+FIXTURE_SERVICE_PORT = "9015"
 
-DEFAULT_SERVER_CONTAINER_NAME2 = "wandb-local-testcontainer2"
-DEFAULT_SERVER_VOLUME2 = "wandb-local-testcontainer-vol2"
+DEFAULT_SERVER_CONTAINER_NAME = "wandb-local-testcontainer"
+DEFAULT_SERVER_VOLUME = "wandb-local-testcontainer-vol"
 
 
 class ConsoleFormatter:
@@ -495,6 +483,106 @@ def pytest_addoption(parser):
     )
 
 
+@dataclasses.dataclass
+class WandbServerSettings:
+    name: str
+    volume: str
+    wandb_server_pull: str
+    wandb_server_image_registry: str
+    wandb_server_image_repository: str
+    wandb_server_tag: str
+    # spin up the server or connect to an existing one
+    wandb_server_use_existing: bool
+    # ports exposed to the host
+    local_base_port: str
+    services_api_port: str
+    fixture_service_port: str
+    # ports internal to the container
+    internal_local_base_port: str = LOCAL_BASE_PORT
+    internal_local_services_api_port: str = SERVICES_API_PORT
+    internal_fixture_service_port: str = FIXTURE_SERVICE_PORT
+    url: str = DEFAULT_SERVER_URL
+
+    base_url: Optional[str] = None
+
+    def __post_init__(self):
+        self.base_url = f"{self.url}:{self.local_base_port}"
+
+
+def check_server_health(
+    base_url: str, endpoint: str, num_retries: int = 1, sleep_time: int = 1
+) -> bool:
+    """Check if wandb server is healthy.
+
+    :param base_url:
+    :param num_retries:
+    :param sleep_time:
+    :return:
+    """
+    for _ in range(num_retries):
+        try:
+            response = requests.get(urllib.parse.urljoin(base_url, endpoint))
+            if response.status_code == 200:
+                return True
+            time.sleep(sleep_time)
+        except requests.exceptions.ConnectionError:
+            time.sleep(sleep_time)
+    return False
+
+
+def spin_wandb_server(settings: WandbServerSettings) -> bool:
+    base_url = settings.base_url
+    app_health_endpoint = "healthz"
+    fixture_url = base_url.replace(
+        settings.local_base_port, settings.fixture_service_port
+    )
+    fixture_health_endpoint = "health"
+
+    if settings.wandb_server_use_existing:
+        return check_server_health(base_url=base_url, endpoint=app_health_endpoint)
+
+    if not check_server_health(base_url, app_health_endpoint):
+        command = [
+            "docker",
+            "run",
+            "--pull",
+            settings.wandb_server_pull,
+            "--rm",
+            "-v",
+            f"{settings.volume}:/vol",
+            "-p",
+            f"{settings.local_base_port}:{settings.internal_local_base_port}",
+            "-p",
+            f"{settings.services_api_port}:{settings.internal_local_services_api_port}",
+            "-p",
+            f"{settings.fixture_service_port}:{settings.internal_fixture_service_port}",
+            "-e",
+            "WANDB_ENABLE_TEST_CONTAINER=true",
+            "--name",
+            settings.name,
+            "--platform",
+            "linux/amd64",
+            f"{settings.wandb_server_image_registry}/{settings.wandb_server_image_repository}:{settings.wandb_server_tag}",
+        ]
+        subprocess.Popen(command)
+        # wait for the server to start
+        server_is_up = check_server_health(
+            base_url=base_url, endpoint=app_health_endpoint, num_retries=30
+        )
+        if not server_is_up:
+            return False
+        # check that the fixture service is accessible
+        return check_server_health(
+            base_url=fixture_url,
+            endpoint=fixture_health_endpoint,
+            num_retries=30,
+        )
+
+    return check_server_health(
+        base_url=fixture_url, endpoint=fixture_health_endpoint, num_retries=10
+    )
+
+
 def pytest_configure(config):
     print("Running tests with wandb version:", wandb.__version__)
     print("Configuring wandb server...")
@@ -525,9 +613,22 @@ def pytest_configure(config):
         pytest.exit("Failed to connect to wandb server")
 
     if config.getoption("--wandb-second-server"):
+        # In CI, we use a docker name and the same ports
+        # Locally, we use localhost and different port mappings
+        default_server_url2 = (
+            os.getenv("WANDB_TEST_SERVER_URL2")
+            if os.getenv("CI")
+            else DEFAULT_SERVER_URL
+        )
+        local_base_port2 = LOCAL_BASE_PORT if os.getenv("CI") else "9180"
+        service_api_port2 = SERVICES_API_PORT if os.getenv("CI") else "9183"
+        fixture_service_port2 = FIXTURE_SERVICE_PORT if os.getenv("CI") else "9115"
+        server_container_name2 = "wandb-local-testcontainer2"
+        server_volume2 = "wandb-local-testcontainer-vol2"
+
         settings2 = WandbServerSettings(
-            name=DEFAULT_SERVER_CONTAINER_NAME2,
-            volume=DEFAULT_SERVER_VOLUME2,
+            name=server_container_name2,
+            volume=server_volume2,
             url=default_server_url2,
             local_base_port=local_base_port2,
             services_api_port=service_api_port2,
@@ -547,10 +648,6 @@ def pytest_configure(config):
         )
         config.wandb_server_settings2 = settings2
 
-        # Container spins up separately in CI
-        if os.getenv("CI"):
-            return
-
         success2 = spin_wandb_server(settings2)
         if not success2:
             pytest.exit("Failed to connect to wandb server2")
@@ -558,52 +655,23 @@ def pytest_configure(config):
 
 def pytest_unconfigure(config):
     clean = config.getoption("--wandb-server-clean")
+    second_server = config.getoption("--wandb-second-server")
+
+    server_settings_objs = [config.wandb_server_settings]
+    if second_server:
+        server_settings_objs += [config.wandb_server_settings2]
+
     if clean != "none":
         print("Cleaning up wandb server...")
     if clean in ("container", "all"):
-        print(
-            f"Cleaning up wandb server container ({config.wandb_server_settings.name}) ..."
-        )
-        command = ["docker", "rm", "-f", config.wandb_server_settings.name]
-        subprocess.run(command, check=True)
-    if clean in ("volume", "all"):
-        print(
-            f"Cleaning up wandb server volume ({config.wandb_server_settings.volume}) ..."
-        )
-        command = ["docker", "volume", "rm", config.wandb_server_settings.volume]
-        subprocess.run(command, check=True)
-
-    if os.getenv("CI"):
-        return
-
-    clean = config.getoption("--wandb-server-clean")
-    if clean != "none":
-        print("Cleaning up wandb server...")
-    if clean in ("container", "all"):
-        print(
-            f"Cleaning up wandb server container ({config.wandb_server_settings.name}) ..."
-        )
-        command = ["docker", "rm", "-f", config.wandb_server_settings.name]
-        subprocess.run(command, check=True)
-
-        if config.getoption("--wandb-second-server"):
-            print(
-                f"Cleaning up wandb server container2 ({config.wandb_server_settings2.name}) ..."
-            )
-            command = ["docker", "rm", "-f", config.wandb_server_settings2.name]
+        for server_settings in server_settings_objs:
+            print(f"Cleaning up wandb server container ({server_settings.name}) ...")
+            command = ["docker", "rm", "-f", server_settings.name]
             subprocess.run(command, check=True)
     if clean in ("volume", "all"):
-        print(
-            f"Cleaning up wandb server volume ({config.wandb_server_settings.volume}) ..."
-        )
-        command = ["docker", "volume", "rm", config.wandb_server_settings.volume]
-        subprocess.run(command, check=True)
-
-        if config.getoption("--wandb-second-server"):
-            print(
-                f"Cleaning up wandb server volume2 ({config.wandb_server_settings2.volume}) ..."
-            )
-            command = ["docker", "volume", "rm", config.wandb_server_settings2.volume]
+        for server_settings in server_settings_objs:
+            print(f"Cleaning up wandb server volume ({server_settings.volume}) ...")
+            command = ["docker", "volume", "rm", server_settings.volume]
             subprocess.run(command, check=True)
